backend done

+++ backend/controller/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id})
     res.status(200).json(goals)
 })
 // @desc Set goals
@@ -22,7 +23,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     }
     )
     res.status(200).json(goal)
@@ -40,6 +42,22 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if (!user)
+    {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id)
+    {
+        res.status(401)
+        throw new Error('User not authorized')
+        }
+
     const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
     res.status(200).json(updateGoal)
 } )
@@ -55,7 +73,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+    const user = await User.findById(req.user.id)
 
+    // check for user
+    if (!user)
+    {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id)
+    {
+        res.status(401)
+        throw new Error('User not authorized')
+        }
     await goal.deleteOne({ id: req.params.id })
     res.status(200).json({ id: req.params.id })
 } )
